% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capper_test_betas.R
\docType{data}
\name{Capper_test_betas}
\alias{Capper_test_betas}
\title{Test GBM DNAm data used for testing}
\format{
A matrix with 485512 rows (CpGs) and 10 columns (samples).
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE90496}
}
\usage{
Capper_test_betas
}
\description{
A random subset of the Capper et al. dataset (GSE90496) including 10
glioblastoma (GBM) samples. DNAm data is represented in a data matrix of
beta values, or the fraction of methylated alleles at the given CpG.
}
\keyword{datasets}
