% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_libraries.R
\docType{data}
\name{GBM_libraries}
\alias{GBM_libraries}
\title{Libraries to deconvolve IDH wildtype glioblastoma DNA methylation data}
\format{
A list of data-frames.
}
\usage{
GBM_libraries
}
\description{
Libraries to deconvolve IDH wildtype glioblastoma DNA methylation data
}
\keyword{datasets}
