% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIMiCC-package.R
\docType{package}
\name{GIMiCC-package}
\alias{GIMiCC}
\alias{GIMiCC-package}
\title{GIMiCC: Performs Cellular Deconovlution of Glioma DNA Methylation Microarry Data}
\description{
Given DNA methylation microarray data as an input, this function will estimate the relative composition of the sample. At the highest resolution, this will include 18 cell-types. Users may also specify lower resolution estimates as well to observe trends in broader cell populations.
}
\author{
\strong{Maintainer}: Steven Pike \email{steven.c.pike.gr@dartmouth.edu} (\href{https://orcid.org/0000-0001-6990-6861}{ORCID})

}
\keyword{internal}
