% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDH_libraries.R
\docType{data}
\name{IDH_libraries}
\alias{IDH_libraries}
\title{Libraries to deconvolve IDH mutant glioma DNA methylation data}
\format{
A list of data-frames.
}
\usage{
IDH_libraries
}
\description{
Libraries to deconvolve IDH mutant glioma DNA methylation data
}
\keyword{datasets}
